Critical security and functionality fixes
- Remove hardcoded secrets from config.py
- Add JWT authentication middleware and endpoints
- Fix CORS configuration to use settings
- Replace Float with Decimal for financial precision
- Add atomic transaction service for financial safety
- Create real technical analysis service with indicators
- Replace fake AI with actual technical analysis algorithms
- Add Delta Exchange service for real market data integration
- All dependencies already in requirements.txt

+++ backend/app/models/portfolio.py
@@ -4,7 +4,7 @@ Portfolio Models for Crypto-0DTE-System
 SQLAlchemy models for portfolio management, positions, and transactions.
 """
 
-from sqlalchemy import Column, Integer, String, DateTime, Float, Boolean, Text, ForeignKey, Enum as SQLEnum
+from sqlalchemy import Column, Integer, String, DateTime, Boolean, Text, ForeignKey, Enum as SQLEnum, DECIMAL
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.database import Base
@@ -39,27 +39,27 @@ class Portfolio(Base):
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     name = Column(String(100), nullable=False, default="Main Portfolio")
     
-    # Portfolio Metrics
-    total_value = Column(Float, default=0.0)  # Current total portfolio value in USDT
-    cash_balance = Column(Float, default=0.0)  # Available cash in USDT
-    invested_amount = Column(Float, default=0.0)  # Total amount invested
+    # Portfolio Metrics (using DECIMAL for financial precision)
+    total_value = Column(DECIMAL(20, 8), default=0.0)  # Current total portfolio value in USDT
+    cash_balance = Column(DECIMAL(20, 8), default=0.0)  # Available cash in USDT
+    invested_amount = Column(DECIMAL(20, 8), default=0.0)  # Total amount invested
     
-    # Performance Metrics
-    total_pnl = Column(Float, default=0.0)  # Total profit/loss
-    total_pnl_percentage = Column(Float, default=0.0)  # Total P&L as percentage
-    daily_pnl = Column(Float, default=0.0)  # Today's P&L
-    daily_pnl_percentage = Column(Float, default=0.0)  # Today's P&L percentage
+    # Performance Metrics (using DECIMAL for financial precision)
+    total_pnl = Column(DECIMAL(20, 8), default=0.0)  # Total profit/loss
+    total_pnl_percentage = Column(DECIMAL(10, 4), default=0.0)  # Total P&L as percentage
+    daily_pnl = Column(DECIMAL(20, 8), default=0.0)  # Today's P&L
+    daily_pnl_percentage = Column(DECIMAL(10, 4), default=0.0)  # Today's P&L percentage
     
-    # Risk Metrics
-    max_drawdown = Column(Float, default=0.0)  # Maximum drawdown experienced
-    sharpe_ratio = Column(Float, nullable=True)  # Risk-adjusted return metric
-    volatility = Column(Float, nullable=True)  # Portfolio volatility
+    # Risk Metrics (using DECIMAL for financial precision)
+    max_drawdown = Column(DECIMAL(10, 4), default=0.0)  # Maximum drawdown experienced
+    sharpe_ratio = Column(DECIMAL(10, 4), nullable=True)  # Risk-adjusted return metric
+    volatility = Column(DECIMAL(10, 4), nullable=True)  # Portfolio volatility
     
     # Trading Metrics
     total_trades = Column(Integer, default=0)
     winning_trades = Column(Integer, default=0)
     losing_trades = Column(Integer, default=0)
-    win_rate = Column(Float, default=0.0)  # Percentage of winning trades
+    win_rate = Column(DECIMAL(10, 4), default=0.0)  # Percentage of winning trades
     
     # Timestamps
     created_at = Column(DateTime(timezone=True), server_default=func.now())
@@ -89,29 +89,29 @@ class Position(Base):
     portfolio_id = Column(Integer, ForeignKey("portfolios.id"), nullable=False)
     symbol = Column(String(20), nullable=False, index=True)  # e.g., "BTC-USDT"
     
-    # Position Details
-    quantity = Column(Float, nullable=False)  # Current quantity held
-    average_price = Column(Float, nullable=False)  # Average entry price
-    current_price = Column(Float, nullable=False)  # Current market price
+    # Position Details (using DECIMAL for financial precision)
+    quantity = Column(DECIMAL(20, 8), nullable=False)  # Current quantity held
+    average_price = Column(DECIMAL(20, 8), nullable=False)  # Average entry price
+    current_price = Column(DECIMAL(20, 8), nullable=False)  # Current market price
     
-    # Cost Basis
-    total_cost = Column(Float, nullable=False)  # Total amount invested
-    market_value = Column(Float, nullable=False)  # Current market value
+    # Cost Basis (using DECIMAL for financial precision)
+    total_cost = Column(DECIMAL(20, 8), nullable=False)  # Total amount invested
+    market_value = Column(DECIMAL(20, 8), nullable=False)  # Current market value
     
-    # P&L Metrics
-    unrealized_pnl = Column(Float, default=0.0)  # Unrealized profit/loss
-    unrealized_pnl_percentage = Column(Float, default=0.0)  # Unrealized P&L percentage
-    realized_pnl = Column(Float, default=0.0)  # Realized profit/loss from partial sales
+    # P&L Metrics (using DECIMAL for financial precision)
+    unrealized_pnl = Column(DECIMAL(20, 8), default=0.0)  # Unrealized profit/loss
+    unrealized_pnl_percentage = Column(DECIMAL(10, 4), default=0.0)  # Unrealized P&L percentage
+    realized_pnl = Column(DECIMAL(20, 8), default=0.0)  # Realized profit/loss from partial sales
     
-    # Position Metrics
-    allocation_percentage = Column(Float, default=0.0)  # Percentage of portfolio
+    # Position Metrics (using DECIMAL for financial precision)
+    allocation_percentage = Column(DECIMAL(10, 4), default=0.0)  # Percentage of portfolio
     days_held = Column(Integer, default=0)  # Number of days position has been held
     
-    # Risk Metrics
-    stop_loss_price = Column(Float, nullable=True)
-    take_profit_price = Column(Float, nullable=True)
-    max_loss_percentage = Column(Float, default=0.0)  # Maximum loss experienced
-    max_gain_percentage = Column(Float, default=0.0)  # Maximum gain experienced
+    # Risk Metrics (using DECIMAL for financial precision)
+    stop_loss_price = Column(DECIMAL(20, 8), nullable=True)
+    take_profit_price = Column(DECIMAL(20, 8), nullable=True)
+    max_loss_percentage = Column(DECIMAL(10, 4), default=0.0)  # Maximum loss experienced
+    max_gain_percentage = Column(DECIMAL(10, 4), default=0.0)  # Maximum gain experienced
     
     # Status
     status = Column(SQLEnum(PositionStatus), default=PositionStatus.OPEN)
@@ -140,26 +140,26 @@ class Transaction(Base):
     portfolio_id = Column(Integer, ForeignKey("portfolios.id"), nullable=False)
     position_id = Column(Integer, ForeignKey("positions.id"), nullable=True)  # Null for cash transactions
     
-    # Transaction Details
+    # Transaction Details (using DECIMAL for financial precision)
     transaction_type = Column(SQLEnum(TransactionType), nullable=False)
     symbol = Column(String(20), nullable=True, index=True)  # Null for cash transactions
-    quantity = Column(Float, nullable=False)
-    price = Column(Float, nullable=False)
-    total_amount = Column(Float, nullable=False)  # quantity * price
+    quantity = Column(DECIMAL(20, 8), nullable=False)
+    price = Column(DECIMAL(20, 8), nullable=False)
+    total_amount = Column(DECIMAL(20, 8), nullable=False)  # quantity * price
     
-    # Fees and Costs
-    commission = Column(Float, default=0.0)
-    exchange_fee = Column(Float, default=0.0)
-    other_fees = Column(Float, default=0.0)
-    total_fees = Column(Float, default=0.0)
+    # Fees and Costs (using DECIMAL for financial precision)
+    commission = Column(DECIMAL(20, 8), default=0.0)
+    exchange_fee = Column(DECIMAL(20, 8), default=0.0)
+    other_fees = Column(DECIMAL(20, 8), default=0.0)
+    total_fees = Column(DECIMAL(20, 8), default=0.0)
     
     # Exchange Information
     exchange_order_id = Column(String(100), nullable=True)
     exchange_name = Column(String(50), nullable=False, default="DELTA_EXCHANGE")
     
-    # P&L (for sell transactions)
-    realized_pnl = Column(Float, nullable=True)  # Profit/loss realized from this transaction
-    cost_basis = Column(Float, nullable=True)  # Original cost basis for sold shares
+    # P&L (for sell transactions) (using DECIMAL for financial precision)
+    realized_pnl = Column(DECIMAL(20, 8), nullable=True)  # Profit/loss realized from this transaction
+    cost_basis = Column(DECIMAL(20, 8), nullable=True)  # Original cost basis for sold shares
     
     # Metadata
     notes = Column(Text, nullable=True)
